--- web/lib/consts.ts
@@ -1,10 +1,0 @@
-export enum HTTP_STATUS_CODE {
-    BAD_REQUEST = 400,
-    OK = 200,
-    CREATED = 201,
-    UNAUTHORIZED = 401,
-    FORBIDDEN = 403,
-    NOT_FOUND = 404,
-    METHOD_NOT_ALLOWED = 405,
-    INTERNAL_SERVER_ERROR = 500,
-}
