+++ web/lib/exceptions/DatabaseExceptions.ts
@@ -1,4 +1,4 @@
-import { HTTP_STATUS_CODE } from "../consts";
+import { HTTP_STATUS_CODE } from "../utils/consts";
 
 export abstract class DatabaseException extends Error {
   code: HTTP_STATUS_CODE;
@@ -9,37 +9,49 @@ export abstract class DatabaseException extends Error {
 }
 
 export class GenericDatabaseErrorException extends DatabaseException {
-    constructor(message = "An error has occurred") {
-        super(message, HTTP_STATUS_CODE.INTERNAL_SERVER_ERROR);
-    }
+  constructor(message = "An error has occurred") {
+    super(message, HTTP_STATUS_CODE.INTERNAL_SERVER_ERROR);
+  }
 }
 
 export class PathNotFoundError extends DatabaseException {
-    constructor(path: string) {
-        super(`Path not found: ${path}`, HTTP_STATUS_CODE.INTERNAL_SERVER_ERROR);
-    }
+  constructor(path: string) {
+    super(`Path not found: ${path}`, HTTP_STATUS_CODE.INTERNAL_SERVER_ERROR);
+  }
 }
 
 export class FailedToFetchError extends DatabaseException {
   constructor(docType: string) {
-      super(`Failed to fetch ${docType} documents.`, HTTP_STATUS_CODE.INTERNAL_SERVER_ERROR);
+    super(
+      `Failed to fetch ${docType} documents.`,
+      HTTP_STATUS_CODE.INTERNAL_SERVER_ERROR
+    );
   }
 }
 
 export class FailedToAddError extends DatabaseException {
   constructor(docType: string) {
-      super(`Failed to add ${docType} document.`, HTTP_STATUS_CODE.INTERNAL_SERVER_ERROR);
+    super(
+      `Failed to add ${docType} document.`,
+      HTTP_STATUS_CODE.INTERNAL_SERVER_ERROR
+    );
   }
 }
 
 export class FailedToEditError extends DatabaseException {
   constructor(docType: string) {
-    super(`Failed to edit ${docType} document.`, HTTP_STATUS_CODE.INTERNAL_SERVER_ERROR);
+    super(
+      `Failed to edit ${docType} document.`,
+      HTTP_STATUS_CODE.INTERNAL_SERVER_ERROR
+    );
   }
 }
 
 export class FailedToDeleteError extends DatabaseException {
   constructor(docType: string) {
-    super(`Failed to delete ${docType} document.`, HTTP_STATUS_CODE.INTERNAL_SERVER_ERROR);
+    super(
+      `Failed to delete ${docType} document.`,
+      HTTP_STATUS_CODE.INTERNAL_SERVER_ERROR
+    );
   }
 }
