--- web/pages/consts.ts
@@ -1,2 +1,0 @@
-export const BABIES_TAB = "babies";
-export const CAREGIVERS_TAB = "caregivers";
